+++ Bearded/Bearded.csproj
@@ -58,6 +58,7 @@
   <ItemGroup>
     <Compile Include="GameStatistics.cs" />
     <Compile Include="GameUI\Beardgame.cs" />
+    <Compile Include="GameUI\GameObjectRenderer.cs" />
     <Compile Include="Performance\PerformanceMonitor.cs" />
     <Compile Include="GameUI\GameScreen.cs" />
     <Compile Include="Performance\PerformanceSummary.cs" />
