+++ GameLogic/Utils.fs
@@ -13,7 +13,7 @@ module Utils =
     type OpenTK.Vector2 with
         member this.FromPolar(angle, radius) = Vector2(Mathf.Cos(angle), Mathf.Sin(angle)) * radius
 
-    let private rnd = System.Random()
-    let public rndSingle () = (single)(rnd.NextDouble())
-    let public rndInt maxValue = rnd.Next(maxValue)
+    let private random = System.Random()
+    let public randomSingle () = (single)(random.NextDouble())
+    let public randomInt maxValue = random.Next(maxValue)
